# api/docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Use your Android phone to send and receive SMS messages via a simple programmable API with end-to-end encryption.",
        "title": "httpSMS API Reference",
        "contact": {
            "name": "support@httpsms.com",
            "email": "support@httpsms.com"
        },
        "license": {
            "name": "AGPL-3.0",
            "url": "https://raw.githubusercontent.com/NdoleStudio/http-sms-manager/main/LICENSE"
        },
        "version": "1.0"
    },
    "host": "api.httpsms.com",
    "basePath": "/v1",
    "paths": {
...
        "/messages/send": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Add a new SMS message to be sent by your Android phone",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "Send an SMS message",
                "parameters": [
                    {
                        "description": "Send message request payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.MessageSend"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.BadRequest"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.Unauthorized"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/responses.UnprocessableEntity"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.InternalServerError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entities.Message": {
            "type": "object",
            "properties": {
                "app_id": {
                    "type": "string",
                    "example": "32343a19-da5e-4b1b-a767-3298a73703cb"
                },
                "contact": {
                    "type": "string",
                    "example": "+18005550100"
                },
                "content": {
                    "type": "string",
                    "example": "This is a sample text message"
                },
                "created_at": {
                    "type": "string",
                    "example": "2022-06-05T14:26:02.302718+03:00"
                },
                "delivered_at": {
                    "type": "string",
                    "example": "2022-06-05T14:26:09.527976+03:00"
                },
                "encrypted": {
                    "type": "boolean",
                    "example": false
                },
                "expired_at": {
                    "type": "string",
                    "example": "2022-06-05T14:26:09.527976+03:00"
                },
                "failed_at": {
                    "type": "string",
                    "example": "2022-06-05T14:26:09.527976+03:00"
                },
                "failure_reason": {
                    "type": "string",
                    "example": "UNKNOWN"
                },
                "id": {
                    "type": "string",
                    "example": "32343a19-da5e-4b1b-a767-3298a73703cb"
                },
                "last_attempted_at": {
                    "type": "string",
                    "example": "2022-06-05T14:26:09.527976+03:00"
                },
                "max_send_attempts": {
                    "type": "integer",
                    "example": 1
                },
                "order_timestamp": {
                    "type": "string",
                    "example": "2022-06-05T14:26:09.527976+03:00"
                },
                "owner": {
                    "type": "string",
                    "example": "+18005550199"
                },
                "received_at": {
                    "type": "string",
                    "example": "2022-06-05T14:26:09.527976+03:00"
                },
                "request_id": {
                    "type": "string",
                    "example": "153554b5-ae44-44a0-8f4f-7bbac5657ad4"
                },
                "request_received_at": {
                    "type": "string",
                    "example": "2022-06-05T14:26:01.520828+03:00"
                },
                "scheduled_at": {
                    "type": "string",
                    "example": "2022-06-05T14:26:09.527976+03:00"
                },
                "scheduled_send_time": {
                    "type": "string",
                    "example": "2022-06-05T14:26:09.527976+03:00"
                },
                "send_attempt_count": {
                    "type": "integer",
                    "example": 0
                },
                "send_time": {
                    "description": "SendDuration is the number of nanoseconds from when the request was received until when the mobile phone send the message",
                    "type": "integer",
                    "example": 133414
                },
                "sent_at": {
                    "type": "string",
                    "example": "2022-06-05T14:26:09.527976+03:00"
                },
                "sim": {
                    "description": "SIM is the SIM card to use to send the message\n* SMS1: use the SIM card in slot 1\n* SMS2: use the SIM card in slot 2\n* DEFAULT: used the default communication SIM card",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entities.SIM"
                        }
                    ],
                    "example": "DEFAULT"
                },
                "status": {
                    "type": "string",
                    "example": "pending"
                },
                "type": {
                    "type": "string",
                    "example": "mobile-terminated"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2022-06-05T14:26:10.303278+03:00"
                },
                "user_id": {
                    "type": "string",
                    "example": "WB7DRDWrJZRGbYrv2CKGkqbzvqdC"
                }
            }
        },
...
        "entities.SIM": {
            "type": "string",
            "enum": [
                "SIM1",
                "SIM2"
            ],
            "x-enum-varnames": [
                "SIM1",
                "SIM2"
            ]
        },
...
        "requests.MessageSend": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string",
                    "example": "This is a sample text message"
                },
                "encrypted": {
                    "description": "Encrypted is an optional parameter used to determine if the content is end-to-end encrypted. Make sure to set the encryption key on the httpSMS mobile app",
                    "type": "boolean",
                    "example": false
                },
                "from": {
                    "type": "string",
                    "example": "+18005550199"
                },
                "request_id": {
                    "description": "RequestID is an optional parameter used to track a request from the client's perspective",
                    "type": "string",
                    "example": "153554b5-ae44-44a0-8f4f-7bbac5657ad4"
                },
                "send_at": {
                    "description": "SendAt is an optional parameter used to schedule a message to be sent in the future. The time is considered to be in your profile's local timezone and you can queue messages for up to 20 days (480 hours) in the future.",
                    "type": "string",
                    "example": "2025-12-19T16:39:57-08:00"
                },
                "to": {
                    "type": "string",
                    "example": "+18005550100"
                }
            }
        },
...
        "responses.BadRequest": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string",
                    "example": "The request body is not a valid JSON string"
                },
                "message": {
                    "type": "string",
                    "example": "The request isn't properly formed"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
...
        "responses.InternalServerError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "We ran into an internal error while handling the request."
                },
                "status": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
        "responses.MessageResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/entities.Message"
                },
                "message": {
                    "type": "string",
                    "example": "Request handled successfully"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "responses.Unauthorized": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string",
                    "example": "Make sure your API key is set in the [X-API-Key] header in the request"
                },
                "message": {
                    "type": "string",
                    "example": "You are not authorized to carry out this request."
                },
                "status": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
        "responses.UnprocessableEntity": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "message": {
                    "type": "string",
                    "example": "validation errors while handling request"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "x-api-Key",
            "in": "header"
        }
    }
}